Initial work to move sign package inside cardinal/server
