+++ cardinal/message_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"pkg.world.dev/world-engine/assert"
+	"pkg.world.dev/world-engine/cardinal/server/sign"
 	"pkg.world.dev/world-engine/cardinal/txpool"
-	"pkg.world.dev/world-engine/sign"
 )
 
 type EmptyMsgResult struct{}
