+++ cardinal/server/sign/sign.go
