+++ cardinal/server/sign/sign_test.go
