+++ e2e/tests/nakama/Dockerfile
@@ -5,7 +5,6 @@ WORKDIR /go/src/app
 COPY evm evm
 COPY cardinal cardinal
 COPY rift rift
-COPY sign sign
 
 # Copy the go module files and download the dependencies
 # We do this before copying the rest of the source code to avoid
