+++ go.work
@@ -7,6 +7,5 @@ use (
 	e2e/tests
 	evm
 	relay/nakama
-	sign
 	rift
 )
