+++ relay/nakama/signer/kms.go
@@ -20,7 +20,7 @@ import (
 	"github.com/rotisserie/eris"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
-	"pkg.world.dev/world-engine/sign"
+	"pkg.world.dev/world-engine/cardinal/server/sign"
 )
 
 var _ Signer = &kmsSigner{}
