--- sign/go.mod
@@ -1,21 +1,0 @@
-module pkg.world.dev/world-engine/sign
-
-go 1.22.1
-
-require (
-	github.com/ethereum/go-ethereum v1.13.10
-	github.com/mitchellh/mapstructure v1.5.0
-	github.com/rotisserie/eris v0.5.4
-	gotest.tools/v3 v3.5.1
-)
-
-require (
-	github.com/btcsuite/btcd/btcec/v2 v2.3.2 // indirect
-	github.com/btcsuite/btcd/chaincfg/chainhash v1.0.2 // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.2.0 // indirect
-	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/holiman/uint256 v1.2.4 // indirect
-	golang.org/x/crypto v0.22.0 // indirect
-	golang.org/x/sys v0.19.0 // indirect
-)
